Moved the snippets to a seperate folder
