+++ src/main/snippets/casus_filter.py
@@ -1,0 +1,21 @@
+from django.utils.translation import gettext_lazy as _
+
+from wagtail.snippets.models import register_snippet
+from wagtail.admin.edit_handlers import FieldPanel
+
+from .snippet_base import SnippetBase
+
+
+@register_snippet
+class CasusFilter(SnippetBase):
+    panels = [
+        FieldPanel("name"),
+    ]
+
+    def __str__(self):
+        return self.name
+
+    class Meta:
+        verbose_name = _("CasusFilter")
+        verbose_name_plural = _("CasusFilters")
+        ordering = ["name"]
